requirements.txt
email-validator>=1.1.3
pydantic>=1.9.0
python-dotenv>=0.19.2
